Lombok, Exception Handling
