+++ src/main/java/com/cap/CapAccountApplication.java
@@ -1,0 +1,13 @@
+package com.cap;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CapAccountApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CapAccountApplication.class, args);
+	}
+
+}
