+++ src/main/java/com/cap/dto/AccountRegistration.java
@@ -1,0 +1,23 @@
+package com.cap.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class AccountRegistration {
+
+	
+	private String first_name;
+	private String last_name;
+	private String ssn;
+	private String dob;
+	private String email;
+	private String mobile_number;
+	private String home_address;
+	private String mailing_address;
+	private String account_type;
+	private double min_balance;
+	
+
+
+}
