+++ src/main/java/com/cap/dto/AuthJson.java
@@ -1,0 +1,13 @@
+package com.cap.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class AuthJson {
+
+	private String email;
+	private String token;
+	
+	
+}
