+++ src/main/java/com/cap/dto/ResponseBean.java
@@ -1,0 +1,14 @@
+package com.cap.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class ResponseBean {
+
+	private String token;
+	private String status;
+	private String tokenException;
+	
+	
+}
