--- src/main/java/com/org/cap/dto/AuthJson.java
@@ -1,20 +1,0 @@
-package com.org.cap.dto;
-
-public class AuthJson {
-
-	private String email;
-	private String token;
-	
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	public String getToken() {
-		return token;
-	}
-	public void setToken(String token) {
-		this.token = token;
-	}
-}
