--- src/main/java/com/org/cap/repository/AccountCrudRepository.java
@@ -1,14 +1,0 @@
-package com.org.cap.repository;
-
-import javax.transaction.Transactional;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import com.org.cap.dto.Account;
-
-@Transactional
-@Repository
-public interface AccountCrudRepository 
-  extends CrudRepository<Account, Integer> {
-}
