+++ src/main/resources/application.properties
@@ -4,3 +4,7 @@ spring.datasource.url=jdbc:mysql://localhost:3306/capgemini
 spring.datasource.username=tempuser
 spring.datasource.password=Fidato@00
 
+#spring.datasource.url=jdbc:mysql://us-cdbr-iron-east-02.cleardb.net:3306/ad_d02c61583ce228b
+#spring.datasource.username=b2d9fdecf3b456
+#spring.datasource.password=278dbe1d
+
