+++ src/test/java/com/org/cap/CapAccountApplicationTests.java
@@ -1,16 +1,15 @@
-package com.org.cap;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class CapAccountApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
+/*
+ * package com.org.cap;
+ * 
+ * import org.junit.Test; import org.junit.runner.RunWith; import
+ * org.springframework.boot.test.context.SpringBootTest; import
+ * org.springframework.test.context.junit4.SpringRunner;
+ * 
+ * @RunWith(SpringRunner.class)
+ * 
+ * @SpringBootTest public class CapAccountApplicationTests {
+ * 
+ * @Test public void contextLoads() { }
+ * 
+ * }
+ */
