+++ src/test/java/com/org/cap/controller/AccountRegistrationControllerTest.java
@@ -1,0 +1,19 @@
+package com.org.cap.controller;
+
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import com.cap.controller.AccountRegistrationController;
+
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@WebMvcTest(AccountRegistrationController.class)
+public class AccountRegistrationControllerTest {
+ 
+	/*
+	 * @Autowired private MockMvc mvc;
+	 */
+ 
+}
